+++ src/yaplox/interpreter.py
@@ -1,17 +1,21 @@
-from yaplox.expr import Binary, Expr, ExprVisitor, Grouping, Literal, Unary
+from yaplox.expr import Any, Binary, Expr, ExprVisitor, Grouping, Literal, Unary
+from yaplox.token import Token
 from yaplox.token_type import TokenType
+from yaplox.yaplox_runtime_error import YaploxRuntimeError
 
 
 class Interpreter(ExprVisitor):
     @staticmethod
-    def _binary_plus(left, right):
+    def _binary_plus(expr, left, right):
         if isinstance(left, (float, int)) and isinstance(right, (float, int)):
             return left + right
 
         if isinstance(left, str) and isinstance(right, str):
             return str(left + right)
 
-        raise ValueError(f"Unknown operands {left} and/or {right}")
+        raise YaploxRuntimeError(
+            expr.operator, "Operands must be two numbers or two strings"
+        )
 
     @staticmethod
     def _is_equal(a, b) -> bool:
@@ -26,6 +30,22 @@ class Interpreter(ExprVisitor):
     def visit_binary_expr(self, expr: Binary):
         left = self._evaluate(expr.left)
         right = self._evaluate(expr.right)
+        token_type = expr.operator.token_type
+
+        # Validate that for the following Tokens the operands are numeric.
+        # Orginal jpox does this in a switch statement. Since python does not
+        # have this statement, the dict method is chosen. To not duplicate this line
+        # over and over, the check is done seperately.
+        if token_type in (
+            TokenType.GREATER,
+            TokenType.GREATER_EQUAL,
+            TokenType.LESS,
+            TokenType.LESS_EQUAL,
+            TokenType.MINUS,
+            TokenType.SLASH,
+            TokenType.STAR,
+        ):
+            self._check_number_operands(expr.operator, left, right)
 
         choices = {
             # Comparison operators
@@ -40,12 +60,18 @@ class Interpreter(ExprVisitor):
             TokenType.MINUS: lambda: float(left) - float(right),
             TokenType.SLASH: lambda: float(left) / float(right),
             TokenType.STAR: lambda: float(left) * float(right),
-            TokenType.PLUS: lambda: self._binary_plus(left, right),
+            TokenType.PLUS: lambda: self._binary_plus(expr, left, right),
         }
-        option = choices[expr.operator.token_type]
-        result = option()
 
-        return result
+        try:
+            option = choices[token_type]
+            result = option()
+            return result
+
+        except KeyError:
+            raise YaploxRuntimeError(
+                expr.operator, f"Unknown operator {expr.operator.lexeme}"
+            )
 
     def visit_grouping_expr(self, expr: Grouping):
         return self._evaluate(expr.expression)
@@ -56,15 +82,24 @@ class Interpreter(ExprVisitor):
     def visit_unary_expr(self, expr: Unary):
         right = self._evaluate(expr.right)
 
-        choices = {
-            TokenType.MINUS: lambda slf: -float(right),
-            TokenType.BANG: lambda slf: not Interpreter._is_truthy(right),
-        }
+        token_type = expr.operator.token_type
+        if token_type == TokenType.MINUS:
+            self._check_number_operand(expr.operator, right)
+            return -float(right)
+        elif token_type == TokenType.BANG:
+            return not Interpreter._is_truthy(right)
 
-        option = choices[expr.operator.token_type]
-        result = option(self)
+    @staticmethod
+    def _check_number_operand(operator: Token, operand: Any):
+        if isinstance(operand, (float, int)):
+            return
+        raise YaploxRuntimeError(operator, f"{operand} must be a number.")
 
-        return result
+    @staticmethod
+    def _check_number_operands(operator: Token, left: Any, right: Any):
+        if isinstance(left, (float, int)) and isinstance(right, (float, int)):
+            return
+        raise YaploxRuntimeError(operator, "Operands must be numbers.")
 
     @staticmethod
     def _is_truthy(obj):